fixed a bug
