+++ server/controllers/saveCoverletter.js
@@ -7,7 +7,7 @@ MongoClient = require("mongodb").MongoClient;
 const { StatusCodes } = require("http-status-codes");
 const User = require("../models/User");
 const CoverLetter = require("../models/coverletter");
-const { BadRequestError,UnauthenticatedError } = require("../errors");
+
 
 uploadCoverLetter = async (req, res) => {
 	MongoClient.connect(process.env.MONGO_URI).then((client) => {
@@ -53,34 +53,9 @@ const saveCoverletter = async (req, res) => {
 		.json({ message: "Cover Letter Saved successfully", coverletter });
 };
 
-const updateCoverLetter = async (req,res)=>{
-	try {
-	const user_id = req.user.userId
-	
 
-	let coverletter = await CoverLetter.findById(req.params.id)
-	
-
-	if(!coverletter){
-		throw new BadRequestError("No cover letter found")
-	}
-	
-	if(coverletter.user_id !=user_id){
-		throw new UnauthenticatedError("you cannot carry out this operation")
-	}
-	
-	coverletter = await CoverLetter.findByIdAndUpdate(req.params.id,req.body,{new:true})
-
-	return res.status(StatusCodes.OK).json({success:true,coverletter})
-	} catch (error) {
-		return res.status(400).json({success:false,error:error.message})
-	}
-	
-
-}
 
 module.exports = {
 	uploadCoverLetter,
-	saveCoverletter,
-	updateCoverLetter
+	saveCoverletter
 };
