+++ server/routes/coverLetterRoutes.js
@@ -6,6 +6,7 @@ const {
 	getAllCoverLettersByAUser,
 	getACoverLetter,
 	deleteCoverLetter,
+	updateCoverLetter
 } = require("../controllers/coverLetterController");
 
 const auth = require("../middleware/authentication");
@@ -17,5 +18,6 @@ router.get("/coverLetter/:id", auth, getACoverLetter);
 router.get("/template", auth, getAllCoverLettersByAUser);
 router.delete("/coverLetter/:id", auth, deleteCoverLetter);
 router.patch("/template/:id", auth, editACoverLetter);
+router.put("/updatecoverletter/:id",auth,updateCoverLetter)
 
 module.exports = router;
