+++ server/routes/cvToCoverLetterRoutes.js
@@ -3,7 +3,7 @@ const router = express.Router();
 const { cvUpload } = require("../controllers/cvUploadController");
 const { sendCoverLetter } = require("../controllers/coverLetterGenerator");
 const { uploadCoverLetter } = require("../controllers/saveCoverletter");
-const { saveCoverletter,updateCoverLetter } = require("../controllers/saveCoverletter");
+const { saveCoverletter} = require("../controllers/saveCoverletter");
 const auth = require("../middleware/authentication");
 const { getACoverLetter } = require("../controllers/coverLetterController");
 
@@ -11,6 +11,6 @@ router.post("/upload", cvUpload);
 router.post("/generate", sendCoverLetter);
 router.post("/coverletterUpload", uploadCoverLetter);
 router.post("/saveCoverletter", auth, saveCoverletter);
-router.put("/updatecoverletter/:id",auth,updateCoverLetter)
+
 
 module.exports = router;
